split magistrate question
